Update download.md
Creat a form page of an open source project tha do't requair a link address. Server or host domain simply oprate at locall device but active like a richtext offline editor.
The form;- is a vcard editor interface for managing files and documents, that detail visible and open on click, able to upload, export,import and share at locall.
The admin;- can edit, upload,import,export,share,allow user to registar and can modify the page.

The admin team;- that will appllied by the admin will have a permission to manage and act, do activities on the title of their position. And the admin team have a profile that include the information, (Name, Age, Sex, position or title,email address, Region,Zone, Woreda, and Kebele)

The form structure;- 
The logo;- desplay on the top title bar of the form the name and logo of the page will be upload

 1. Login form; is attached to the top of the page, that can oprate in offline state like a groove account or database and the subscriber will have access to the page to view,comment and give feedback to the admin
2. Search engine; a navigation bar on the top mid of the form to type and search locall fiels and information from the page form.
3. Placeholder;- dropout a list of task force as active plain text (Industry extension, Job creation, Institution capacity building, Marketing, Planning)
4. Menu ;-  a button represented by a three row at the left corner of the page a place holder and vertical scroll  containing  a list of items, (about,the team profile and contacts.)
5. Contents;- a horizontal scroll like slide containing a block of ( Data, Plan, Report, Checklist, Feedback, Guide,and Gallary) each block consists a list of item that dropout on click. 
    The Data block;- dropout a list of items on click ( Basic, Demography, Fisical)
   The Plan block;- dropout a list of items on click ( annual, strategic,BSC, seasonal)
  The Report block;- dropout a list of items on click (monthly, weekly, annual, seasonal)
  The Checklist block;- dropout a list of items on click (For customer, from client)
  The Guied block;- dropout a list of items on click ( rules, regulation, scientific research, other)
 
 The Galleryblock;- dropout a list of items on click ( image video)
Dashboard; a place to edit and publish tasks desplay and view activities 
Editor tools; insert basic word and image editor tools
The footer of the page consist the name and address of the organization, as Woliyusite, Amhara Region, Oromo Zone, Dewe Harewa, Bora, Ethiopia and syncornise and insert the  link address associated with;- kemalmulatu19@gmail.com